pyproject.toml
[project]
name = "pymtech-core"
version = "0.1.1"
description = "Pymtechbasic core library for settings, loggers, etc"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-settings>=2.2.0",
    "boto3>=1.35.14",
    "orjson>=3.10.7",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2",
    "ruff>=0.6.4",
]
